Fixes: OPENSTACK-61
Problem: Default value for boolean mgmt floating ip was set to false

Analysis: When no floating IP is given to the mgmt interface on the VE,
the onboard server must have a route to that subnet. Having a floating
IP ensures this IP is publicly routable (in our lab at least), so this
variable was set to a default value of true.

Unit/Robot tests: