Fixes: Initial import of gi-simplify heat templates
Problem:
The infrastructure for gi-simplify was previously only
available in hype. This commit makes it available to OpenStack
Kilo clouds in general

Analysis:
This is part of a larger effort to make available to the SP community
a set of OpenStack heat templates that can be used in conjunction with
configuration management tools to create P.O.C and reference implementations
of common SP feature requests. This particular solution emphasizes CGNAT,
PEM, and several other use cases

Unit/Robot tests:
Using the heat client, this was tested on the Boulder OpenStack install
which, at the time of this commit, is based on OpenStack Kilo