Addresses some of the issues regarding heat resources in non-public locations
Fixes: Issue #16

Problem: All templates need to reference proper lcoation of yaml
resources

Analysis: Fixed many templates to refer to this repo, when possible, as
to make the resources needed by the templates found on github. The
remaining problem with this technique is that this repo is still
private. Once it is public, we'll reap the benefits. The outstanding
issue for the clustering templates, is that it relies upon the onboard
and odk packages of old, which do not exist anywhere publicly
accessible. I am creating a ticket for this seperately.

Unit/Robot tests: