Functional tests needed for patch_upload_ve_image
Issues: Fixes Issue #50

Problem: Need functional tests for this template to be able to bump it
to supported

Analysis: Three functional tests added, one for each VE image version.
The stack is created and once it is complete, we verify that the VE
image has been uploaded to glance.

Tests: All tests run manually and passed