Functional tests need for ve image boot template
Issues: Issue #51

Problem: Testing for common/f5_ve_standalone_3_nic.yaml template for
each VE image supported.

Analysis: Implemented three tests, one for each supported VE image,
12.0, 11.6.0, and 11.5.4. The fixtures used for each test are a
combination of those in the conftest in this repo and the
f5-openstack-test fitures that are pip installed.

Tests: The 11.5.4 test fails, as we suspected, but when we allow the
startup script to be overriden in the f5-openstack-image-prep tool, we
will be able to use the old startup script to fix this. We will also
address this failure in that repo.