Patched VE images torn down even though no-teardown flag is given to pytest
Issues: Issue #53

Problem: When running a functional test for patching a VE image, a user
can provide the '--no-teardown-glance-images' to prevent the removal of
those patched VE images once the test completes. This was not working,
and it was due to a bad setting of the config option.

Analysis: Fixed the config option setting by making the default action
as store true, and the default value of this option as store false. Also
made these tests a little more flexible by not hardcoding the test
template location in the suite, but by walking back up the directory
tree from the conftest.py location and getting access to the unsupported
directory.

Tests: Ran the tests both with and without this setting. It works as
expected. Added back in the apt-cache proxy url.