Functional test needed for launching the deploy_lb template
Issues:
Fixes #63

Problem:
We would like to promote this template in unsupported/f5_plugins/ to
supported. This issue is to track that movement and the tests to
validate this template.

Analysis:
Created a functional test and moved the template from unsupported to
supported. The template itself validates that the stack is being created
properly. This is done with wait conditions. As long as those are met,
the stack will be create complete. Once that happens, we can be
guaranteed that the iapp is functioning as it should, the backend server
is up and running, and the client is up (because it makes a request to
the VIP).

Tests:
This test passes.