Add tests for save and sync plugins
Issues:
Fixes #102

Problem:
New plugins are being added to f5-openstack-heat-plugins for
F5::Sys::Save and F5::Cm::Sync and they need to be added to an
f5_supported template for testing.

Analysis:
Added use of plugins in active_standby template to save config on both
devices and sync config on one device to the group.

Tests:
All tests pass