Patch image template relies too much upon apt-get installing latest packages
Issues:
Fixes #140

Problem:
In the patch_upload_ve_image.yaml supported heat template, we do an
apt-get install of several packages from ubuntu. These are from latest,
and we can't guarantee that these packages will not change. That is a
bad thing if we want this template to continue working. This is
especially important with the python-keystoneclient and
python-glanceclient. They are required by the f5-openstack-image-prep
repository. That repo now has a setup.py with these packages pinned,
therefore we don't need it in the heat template here. We'll pin what we
can in the this template and remove the keystone and glance clients.

Analysis:
Edited the bash script in the Heat template to remove the keystone and
glance installations. Also pinned the other packages to specific
versions so they won't move.

Tests:
Tested by patching a 12.1.1 image and launching it in my local liberty
OpenStack. Logged into the GUI after the image was launched.